chunkserver, master: Increase resolution of the MA<->CS timeout
Increase resolution of master<->chunskerver connection timeout to reduce
latency of chunkserver failure detection and recovery.

+++ src/common/cstoma_communication.h
@@ -26,7 +26,7 @@ LIZARDFS_DEFINE_PACKET_SERIALIZATION(
 		cstoma, registerHost, LIZ_CSTOMA_REGISTER_HOST, 0,
 		uint32_t, ip,
 		uint16_t, port,
-		uint16_t, timeout,
+		uint32_t, timeout,
 		uint32_t, csVersion)
 
 LIZARDFS_DEFINE_PACKET_SERIALIZATION(
